Rename package and refactor most functions
Because of the definition of "mark" in Emacs,`org-marking' was much too
confusing a name for this package, so rename to `org-grading'.
Meanwhile, refine the headline-generation code, import some of hte
supporting functions from my config files and rename, begin some code
cleanup.

Still lots left to do, including removing some cruft, and getting
org-generate-tables to work agian with org 9.0.